Disable mocha CLI npm tasks that depend on coverage or testing
The reason is that in the current dev environment configuration, there
is/are incompatabilites with DOM modules. The following all experience
issues, regardless of whether using nvm v0.12 or v4.0:

	* phantomJS
	* jsdom (inv v3)
	* jsdom-global
	* jsdom-compat
	* jquery

which leaves window.document undefined, along with the corresponding
DOM API (querySelector et al), and guarantees assertion failure.

Disabling coverage, pre-commit and watcher tasks should prevent the
errors from being displayed. Most development from this point on is
focused on the DOM.